+++ gulpfile.js
@@ -7,6 +7,10 @@ const hash = require("gulp-hash");
 const sass = require("gulp-sass");
 const size = require("gulp-size");
 const autoprefixer = require("gulp-autoprefixer");
+const imagemin = require("gulp-imagemin");
+const imageminMozjpeg = require("imagemin-mozjpeg");
+const imageminPngquant = require("imagemin-pngquant");
+const imageminSvgo = require("imagemin-svgo");
 
 // internal gulp plugins
 const metalsmith = require("./gulp/metalsmith");
@@ -23,15 +27,6 @@ const paths = config.paths;
 
 // Placeholder for the environment sniffing variable
 const PRODUCTION = false;
-
-// Copy all images
-// If in PRODUCTION perform some magic
-function images(done) {
-  return gulp
-    .src(paths.images.src, { dot: true })
-    .pipe(gulp.dest(paths.images.dest))
-    .pipe(size());
-}
 
 const sassOptions = {
   eyeglass: {}
@@ -73,13 +68,26 @@ function clean(done) {
   done();
 }
 
+function imageOptim() {
+  return gulp
+    .src(paths.images.src)
+    .pipe(
+      imagemin([
+        imageminMozjpeg({ quality: 85 }),
+        imageminPngquant({ quality: "65-80" }),
+        imageminSvgo({ plugins: [{ removeViewBox: false }] })
+      ])
+    )
+    .pipe(gulp.dest(paths.images.dest));
+}
+
 function watch(done) {
   gulp.watch(
     paths.scripts.modules,
     gulp.series(scripts.bundle, browserSync.reload)
   );
   gulp.watch(paths.styles.src, gulp.series(styles, browserSync.reloadCSS));
-  gulp.watch(paths.images.src, gulp.series(images, browserSync.reload));
+  gulp.watch(paths.images.src, gulp.series(imageOptim, browserSync.reload));
   gulp.watch(paths.assets.src, gulp.series(assets, browserSync.reload));
   gulp.watch(
     [paths.pages.src, paths.templates.src],
@@ -92,7 +100,13 @@ function watch(done) {
 gulp.task(
   "build",
 
-  gulp.parallel(assets, styles, scripts.bundle, images, metalsmith.build)
+  gulp.parallel(
+    assets,
+    imageOptim,
+    styles,
+    scripts.bundle,
+    metalsmith.build
+  )
 );
 
 gulp.task("default", gulp.series("build", browserSync.initTask, watch));